==> 3_TP_4_kModes .ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.cluster import DBSCAN, KMeans

Cargamos dataset preprocesado

# Levanta datos prepocesados
data_cat = pd.read_csv("data_cat.csv")
data_nom = pd.read_csv("data_nom.csv")
data_lan = pd.read_csv("data_nom.csv")

onehot = OneHotEncoder(sparse=False)

onehotfit = onehot.fit(data_cat)
onehotfit

data_cat_dummies = onehot.transform(data_cat)

data_cat_dummies = pd.DataFrame(data_cat_dummies, columns=onehot.get_feature_names(data_cat.columns))

onehot.get_feature_names(data_cat.columns)

### Ahora vamos a standarizar las variables nominales

scaler = StandardScaler()
data_nom_scaled = scaler.fit_transform(data_nom)
data_nom_scaled = pd.DataFrame(data_nom_scaled, columns=data_nom.columns)
data_nom_scaled

data_nom_cat = pd.merge(data_nom_scaled,data_cat_dummies,how="inner", left_index=True, right_index=True)
data_nom_cat_analisis = data_nom_cat.iloc[:,0:13]
data_nom_scaled_analisis = pd.merge(data_nom_scaled,data.iloc[:,1:4],how="inner", left_index=True, right_index=True)


kmeans_model = KMeans()
model_fit = kmeans_model.fit(data_nom_cat_analisis)

kmeans_model = KMeans()
model_fit = kmeans_model.fit(data_nom)

data_nom

model_fit.n_clusters

sns.pairplot(data_nom_cat_analisis)

from sklearn.metrics import silhouette_score,silhouette_samples

sil=[]
k_values=range(2,8);

for k in k_values:
    kmeans=KMeans(n_clusters=k,n_init=10,random_state=0) # Numero de repeticiones=10
    kmeans.fit(data_nom_cat_analisis)
    score=silhouette_score(data_nom_cat_analisis,kmeans.labels_)
    sil.append(score)

sns.lineplot(x=k_values,y=sil,marker='o',size=30,legend=False);
plt.ylabel('Silhouette score',fontsize=15);plt.xlabel('Número de clusters',fontsize=15);

from sklearn.cluster import KMeans


X=data_nom_cat_analisis

scaler=StandardScaler()

X_scaled=scaler.fit_transform(X);

kmeans=KMeans(n_clusters=4,n_init=10,random_state=0) # Numero de repeticiones=10
kmeans.fit(X_scaled)

labels = kmeans.labels_ # Etiquetas que indican a qué cluster pertenecen los datos
centroids = kmeans.cluster_centers_ # Coordenadas de los centroides

sns.scatterplot(x=X_scaled[:,0],y=X_scaled[:,1],hue=labels,legend='full');
plt.xlabel('Income',fontsize=15);plt.ylabel('Spending Score',fontsize=15);
plt.scatter(centroids[:,0],centroids[:,1],marker='x',s=50,color='k');

sns.scatterplot(data=data_nom_scaled, x='age', y='income',hue='height')

data_nom_height_mask = (data_nom['height']>45) & (data_nom['income']>0) & (data_nom['income']<250000)
data_nom_limp = data_nom[data_nom_height_mask]
data_nom_limp

sns.pairplot(data_nom_limp, hue='sex')

sns.pairplot(data_nom_scaled_analisis, hue='sex')

sns.pairplot(data_nom_scaled_analisis, hue='status')

sns.pairplot(data_nom_scaled_analisis, hue='orientation')

from sklearn.metrics import silhouette_score,silhouette_samples

sil=[]
k_values=range(2,12);

for k in k_values:
    kmeans=KMeans(n_clusters=k,n_init=10,random_state=0) # Numero de repeticiones=10
    kmeans.fit(data_nom_limp[['age','height']])
    score=silhouette_score(data_nom_limp[['income','height']],kmeans.labels_)
    sil.append(score)

sns.lineplot(x=k_values,y=sil,marker='o',size=30,legend=False);
plt.ylabel('Silhouette score',fontsize=15);plt.xlabel('Número de clusters',fontsize=15);

from sklearn.cluster import KMeans


X=data_nom_limp[['age','height']]

scaler=StandardScaler()

X_scaled=scaler.fit_transform(X);

kmeans=KMeans(n_clusters=2,n_init=10,random_state=0) # Numero de repeticiones=10
kmeans.fit(X_scaled)

labels = kmeans.labels_ # Etiquetas que indican a qué cluster pertenecen los datos
centroids = kmeans.cluster_centers_ # Coordenadas de los centroides

sns.scatterplot(x=X_scaled[:,0],y=X_scaled[:,1],hue=labels,legend='full');
plt.xlabel('Income',fontsize=15);plt.ylabel('Height',fontsize=15);
plt.scatter(centroids[:,0],centroids[:,1],marker='x',s=50,color='k');

# Separa en dos clusters, pero no de la fomra correcta

sns.pairplot(data_nom_scaled)

from sklearn.metrics import silhouette_score,silhouette_samples

sil=[]
k_values=range(2,8);

for k in k_values:
    kmeans=KMeans(n_clusters=k,n_init=10,random_state=0) # Numero de repeticiones=10
    kmeans.fit(data_nom_scaled)
    score=silhouette_score(data_nom_scaled_analisis[['age','height']],kmeans.labels_)
    sil.append(score)

sns.lineplot(x=k_values,y=sil,marker='o',size=30,legend=False);
plt.ylabel('Silhouette score',fontsize=15);plt.xlabel('Número de clusters',fontsize=15);

from sklearn.cluster import KMeans


X_1=data_nom_scaled_analisis[['age','height']]

scaler=StandardScaler()

X_scaled_1=scaler.fit_transform(X);

kmeans=KMeans(n_clusters=2,n_init=10,random_state=0) # Numero de repeticiones=10
kmeans.fit(X_scaled)

labels = kmeans.labels_ # Etiquetas que indican a qué cluster pertenecen los datos
centroids = kmeans.cluster_centers_ # Coordenadas de los centroides

sns.scatterplot(x=X_scaled_1[:,0],y=X_scaled_1[:,1],hue=labels,legend='full');
plt.xlabel('Age',fontsize=15);plt.ylabel('Height',fontsize=15);
plt.scatter(centroids[:,0],centroids[:,1],marker='x',s=50,color='k');

dbsscan_model = DBSCAN()
model2_fit = dbsscan_model.fit(data_nom_cat)

model2_fit.core_sample_indices_

from scipy.spatial.distance import pdist,squareform

X=data_nom_cat
minPts=5;
k=minPts-1;

D=squareform(pdist(X));

k_distances=np.zeros(D.shape[0]);

for i in range(D.shape[0]):
    distances=np.sort(D[i]);
    k_distances[i]=distances[k];
k_distances=np.sort(k_distances);
k_distances=k_distances[::-1];
    
plt.plot(k_distances);
plt.xlabel('Rank');
plt.ylabel('K-distance');
plt.hlines([0.2,0.15],0,400,linestyles='dashed');

# Modelado con K-Models

Utilización de K-Models para análisis de variables categóricas

# Se realiza una copia de dataset

data_cat_copy = data_cat.copy()
data_cat_copy.head(5)

# Preparación de dataframe
from sklearn import preprocessing
le = preprocessing.LabelEncoder()
data_cat_copy = data_cat_copy.apply(le.fit_transform)
data_cat_copy.head(10)

data_cat_copy.status.value_counts()

data_cat.status.value_counts()

cost = []
for num_clusters in list(range(1,201,10)):
    kmode = KModes(n_clusters=num_clusters, init = "Cao", n_init = 1, verbose=1)
    kmode.fit_predict(data_cat_copy)
    cost.append(kmode.cost_)

y = np.array([i for i in range(1,201,10)])
plt.plot(y,cost)

# Se eligen 8 clusters
km_cao = KModes(n_clusters=8, init = "Cao", n_init = 1, verbose=1)
fitClusters_cao = km_cao.fit_predict(data_cat_copy)

fitClusters_cao

clusterCentroidsDfsDf = pd.DataFrame(km_cao.cluster_centroids_)
clusterCentroidsDf.columns = data_cat_copy.columns

Se combinan los cluster predichos con el DF original

data_cat_copy = data_cat.reset_index()


clustersDf = pd.DataFrame(fitClusters_cao)
clustersDf.columns = ['cluster_predicho']
combinadoDf = pd.concat([data_cat_copy, clustersDf], axis = 1).reset_index()
combinadoDf = combinadoDf.drop(['index', 'level_0'], axis = 1)

Identificación de clusters

cluster_0 = combinadoDf[combinadoDf['cluster_predicho'] == 0]
cluster_1 = combinadoDf[combinadoDf['cluster_predicho'] == 1]
cluster_2 = combinadoDf[combinadoDf['cluster_predicho'] == 2]
cluster_3 = combinadoDf[combinadoDf['cluster_predicho'] == 3]
cluster_4 = combinadoDf[combinadoDf['cluster_predicho'] == 4]
cluster_5 = combinadoDf[combinadoDf['cluster_predicho'] == 5]
cluster_6 = combinadoDf[combinadoDf['cluster_predicho'] == 6]
cluster_7 = combinadoDf[combinadoDf['cluster_predicho'] == 7]


plt.subplots(figsize = (15,5))
sns.countplot(x=combinadoDf['orientation'],order=combinadoDf['orientation'].value_counts().index,hue=combinadoDf['cluster_predicho'])
plt.show()

plt.subplots(figsize = (15,5))
sns.countplot(x=combinadoDf['drinks'],order=combinadoDf['drinks'].value_counts().index,hue=combinadoDf['cluster_predicho'])
plt.show()

plt.subplots(figsize = (15,5))
sns.countplot(x=combinadoDf['smokes'],order=combinadoDf['smokes'].value_counts().index,hue=combinadoDf['cluster_predicho'])
plt.show()
