==> 5_TP_4_cupidus_.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.cluster import DBSCAN, KMeans
from numpy import dot
from numpy.linalg import norm

data_cat = pd.read_csv("data_cat.csv")
data_nom = pd.read_csv("data_nom.csv")
data_lan = pd.read_csv("data_nom.csv")

# removemos income por la baja cantidad de datos. no se si quedamos en eso, de u;ltoima que vuelva
data_nom.drop("income", axis = 1)

# Vamos a usar datos normalisados para los nominales
scaler = StandardScaler()
data_nom_scaled = scaler.fit_transform(data_nom)
data_nom_scaled = pd.DataFrame(data_nom_scaled, columns=data_nom.columns)
data_nom_scaled

# La idea es no incluir sexo y orientacion en el calculo de distancia: los guardamos separado para no incidir en las agrupaciones
# pregunta y "estado"? si lo dejamos en distancia va a juntar solteros con solteros y casados con casados verque onda

data_sex = data_cat.pop("sex")
data_ori = data_cat.pop("orientation")

onehot = OneHotEncoder(sparse=False)
onehotfit = onehot.fit(data_cat)
data_cat_dummies = onehot.transform(data_cat)
data_cat_dummies = pd.DataFrame(data_cat_dummies, columns=onehot.get_feature_names(data_cat.columns))
data_cat_dummies

#uno los dats en un solo vector por persona:

data_all = pd.merge(data_nom_scaled, data_cat_dummies,how="inner", left_index=True, right_index=True)
data_all = pd.merge(data_all,data_lan,how="inner", left_index=True, right_index=True)

# defino una function de distancia entre personas: 

def distanciaEntrePersonas(persona1, persona2):
    # segun google esto deberia andar. no se como chequearlo. por lo pronto corre
    return np.dot(persona1, persona2) / (np.linalg.norm(persona1) * np.linalg.norm(persona2))

#busco los mejores N candidatos:

numero_recomendaciones = 10 #para tirar un corte y que no me recomiende a todes :P 

recomendaciones_x_cliente = []

for cliente in data_all:

    scores = [] # lista de scores

    for idx, persona in data_all.iterrows():

        dist = distanciaEntrePersonas(cliente, persona)
        scores.append((idx, dist))  # agrego la tupla idx, dist a una lista

    #ahora aca tengo que ordenar la lista scores en orden ascendente de distancia:

    scores.sort(key = lambda x: x[1]) #estoy ordenanjdo por el segundo elem,ento de la tupla q es la dist again, segun google y le creo

    print (scores)

    idx_ofertas = []
    # recorrer scores y devolver las recomendaciones que complan con criterios de orientacion y sexo

    for idx, score in scores:

        if cliente.sex #noinono -.- :
            idx_ofertas.append(idx)

        if len(idx_ofertas) >= numero_recomendaciones:
            break;

    # en idx_ofertas estan los indices de los mejores
    recomendaciones_x_cliente.append(idx_ofertas)
